+++ CHANGELOG.md
@@ -6,5 +6,6 @@ Upcoming
 
 ### Breaking changes
 
+* Eliminate `ClientWithExecutor`, use `Client::create_with_executor()` instead.
 * Eliminate `MemoryClient`. The memory client is now called the emulator and can
   be created with `Client::new_emulator()`.